+++ src/tg/mod.rs
@@ -2,4 +2,5 @@ pub mod accounts;
 pub mod handlers;
 pub mod history;
 pub mod login;
+pub mod media;
 pub mod util;